add Like model, add article and categories policies, update single article in front and update approved comments

+++ app/Http/Controllers/Admin/ArticleController.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Traits\Uploads;
 use App\Models\Category;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ArticleController extends Controller
 {
@@ -108,43 +109,50 @@ class ArticleController extends Controller
      */
     public function update(Request $request, Article $article)
     {
-        $this->validate(request(), [
-            'title' => 'required',
-            'short_description' => 'required',
-            'image' => 'nullable|image',
-            'content' => 'required'
-        ]);
+        $response = Gate::inspect('update', $article);
 
-        $user = auth()->user();
+        if ($response->allowed()) {
+            $this->validate(request(), [
+                'title' => 'required',
+                'short_description' => 'required',
+                'image' => 'nullable|image',
+                'content' => 'required'
+            ]);
 
-        $image = null;
-        $file = null;
-        if($request->has('image')){
-            $image = $this->uploadImage($request->image, Article::class);
-        }else{
-            $image = $article->image;
+            $user = auth()->user();
+
+            $image = null;
+            $file = null;
+            if($request->has('image')){
+                $image = $this->uploadImage($request->image, Article::class);
+            }else{
+                $image = $article->image;
+            }
+
+            if($request->has('zip_files')){
+                $file = $this->uploadeFilesAndZip($request->zip_files, Article::class);
+            }else{
+                $file = $article->file;
+            }
+
+            $data = [
+                'title' => $request->title,
+                'short_description' => $request->short_description,
+                'content' => $request->content,
+                'user_id' => $user->id,
+                'image' => $image,
+                'file' => $file,
+                'approved' => $user->role->id == 1 ? true : false
+            ];
+
+            $article->update($data);
+            $article->categories()->sync($request->categories);
+
+            return redirect()->route('articles.index');
+        } else {
+            return redirect()->back()->withErrors(['error' => $response->message()]);
         }
 
-        if($request->has('zip_files')){
-            $file = $this->uploadeFilesAndZip($request->zip_files, Article::class);
-        }else{
-            $file = $article->file;
-        }
-
-        $data = [
-            'title' => $request->title,
-            'short_description' => $request->short_description,
-            'content' => $request->content,
-            'user_id' => $user->id,
-            'image' => $image,
-            'file' => $file,
-            'approved' => $user->role->id == 1 ? true : false
-        ];
-
-        $article->update($data);
-        $article->categories()->sync($request->categories);
-
-        return redirect()->route('articles.index');
     }
 
     /**
@@ -155,22 +163,38 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
-        if($article->image){
-            $imagePath = public_path($article->image);
-            if (file_exists($imagePath)) {
-                unlink($imagePath);
+        $response = Gate::inspect('delete', $article);
+        if ($response->allowed()) {
+
+            if($article->image){
+                $imagePath = public_path($article->image);
+                if (file_exists($imagePath)) {
+                    unlink($imagePath);
+                }
             }
-        }
-        if($article->file){
-            $filePath = public_path($article->file);
-            if (file_exists($filePath)) {
-                unlink($filePath);
+            if($article->file){
+                $filePath = public_path($article->file);
+                if (file_exists($filePath)) {
+                    unlink($filePath);
+                }
             }
+            $article->delete();
+            return redirect()->route('articles.index');
+
+        }else{
+
+            return redirect()->back()->withErrors(['error' => $response->message()]);
+
         }
-        $article->delete();
-        return redirect()->route('articles.index');
     }
 
+    /**
+     * CKEditor Image save in public storage.
+     *
+     * @param Request $request
+     *
+     * @return string
+     */
     public function attachImage(Request $request)
     {
         $this->validate(request(), [
